Solution that works
Not the best, a recursive function is far better but this was my initial idea. Convert the numbers to strings, reverse the strings, convert to ints and do the math, then convert back to a string and reverse the string. Then output the listnode.

Keeping everything as numbers appears to be a better solution and using a recursive function.